requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
configparser==3.5.0
cryptography==2.3
Flask==1.0.2
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nltk==3.3
pycparser==2.18
PyMySQL==0.9.2
six==1.11.0
SQLAlchemy==1.2.10
Werkzeug==0.14.1
